Fix network creation when nw already exists

+++ scripts/create_isolated_network.sh
@@ -14,9 +14,13 @@ if [[ "$#" -ne 1 ]]; then
     exit 1
 fi
 
-set -x
 # Read from config
 BR_NAME=${1}
+
+if [[ -f "/etc/libvirt/qemu/networks/${BR_NAME}.xml" ]]; then
+    echo "File found /etc/libvirt/qemu/networks/${BR_NAME}.xml"
+    exit
+fi
 
 if [[ ! -f "/etc/libvirt/qemu/networks/${BR_NAME}.xml" ]]; then
 sudo tee -a /etc/libvirt/qemu/networks/${BR_NAME}.xml > /dev/null << EOF
@@ -25,15 +29,10 @@ sudo tee -a /etc/libvirt/qemu/networks/${BR_NAME}.xml > /dev/null << EOF
     <bridge name='${BR_NAME}' stp='on' delay='0' />
 </network>
 EOF
-fi
 
-if [[ -f "/etc/libvirt/qemu/networks/${BR_NAME}.xml" ]]; then
-    echo "File found /etc/libvirt/qemu/networks/${BR_NAME}.xml"
-    exit
-else
     sudo virsh net-define /etc/libvirt/qemu/networks/${BR_NAME}.xml
     sudo virsh net-start ${BR_NAME}
     sudo virsh net-autostart ${BR_NAME}
 fi
 
-set +xe
+set +e
